+++ CHANGELOG.rst
@@ -20,6 +20,10 @@ number of the code change for that issue.  These PRs can be viewed at:
 
 3.7.1 (unreleased)
 ======================
+- Exclude single filter images from the generation of the total detection
+  image to minimize cosmic ray contamination, unless there are only single
+  filter images in the visit. [#nnnn]
+
 - Implemented a series of bug fixes for the segmentation catalog [#1793]
 - Define the threshold image to be (nsigma * background_rms).
 - Fixed bug in the generation of the threshold image - ensure the final